+++ anyblok/bloks/anyblok_core/tests/test_core_sql_base.py
@@ -36,147 +36,129 @@ class TestCoreSqlBase:
 
     def test_from_primary_key(self, rollback_registry):
         registry = rollback_registry
-        Model = registry.System.Model
-        model = Model.query().first()
-        model2 = Model.from_primary_keys(name=model.name)
-        assert model == model2
+        Blok = registry.System.Blok
+        blok = Blok.query().first()
+        blok2 = Blok.from_primary_keys(name=blok.name)
+        assert blok == blok2
 
     def test_from_primary_keys(self, rollback_registry):
         registry = rollback_registry
-        Column = registry.System.Column
-        column = Column.query().first()
-        column2 = Column.from_primary_keys(model=column.model, name=column.name)
-        assert column == column2
+        Cache = registry.System.Cache
+        cache = Cache.query().first()
+        cache2 = Cache.from_primary_keys(id=cache.id)
+        assert cache == cache2
 
     def test_get_primary_key(self, rollback_registry):
         registry = rollback_registry
-        assert registry.System.Model.get_primary_keys() == ["name"]
+        assert registry.System.Blok.get_primary_keys() == ["name"]
 
     def test_get_primary_keys(self, rollback_registry):
         registry = rollback_registry
-        pks = registry.System.Column.get_primary_keys()
-        model_in_pks = "model" in pks
-        name_in_pks = "name" in pks
-        assert model_in_pks is True
-        assert name_in_pks is True
+        pks = registry.System.Cache.get_primary_keys()
+        assert "id" in pks
 
     def test_to_primary_key(self, rollback_registry):
         registry = rollback_registry
-        model = registry.System.Model.query().first()
-        assert model.to_primary_keys() == dict(name=model.name)
+        blok = registry.System.Blok.query().first()
+        assert blok.to_primary_keys() == dict(name=blok.name)
 
     def test_to_primary_keys(self, rollback_registry):
         registry = rollback_registry
-        column = registry.System.Column.query().first()
-        assert column.to_primary_keys() == {
-            "model": column.model,
-            "name": column.name,
+        cache = registry.System.Cache.query().first()
+        assert cache.to_primary_keys() == {
+            "id": cache.id,
         }
 
     def test_fields_description(self, rollback_registry):
         registry = rollback_registry
-        Model = registry.System.Model
+        Cache = registry.System.Cache
+        selections = [
+            (k, v.__doc__ and v.__doc__.split("\n")[0] or k)
+            for k, v in registry.loaded_namespaces.items()
+        ]
         res = {
-            "description": {
-                "id": "description",
-                "label": "Description",
+            "id": {
+                "id": "id",
+                "label": "Id",
+                "model": None,
+                "nullable": False,
+                "primary_key": True,
+                "type": "Integer",
+            },
+            "method": {
+                "id": "method",
+                "label": "Method",
+                "model": None,
+                "nullable": False,
+                "primary_key": False,
+                "type": "String",
+            },
+            "registry_name": {
+                "id": "registry_name",
+                "label": "Registry name",
+                "model": None,
+                "nullable": False,
+                "primary_key": False,
+                "type": "ModelSelection",
+                "selections": selections,
+            },
+        }
+        assert Cache.fields_description() == res
+
+    def test_fields_description_limited_field(self, rollback_registry):
+        registry = rollback_registry
+        Blok = registry.System.Blok
+        res = {
+            "installed_version": {
+                "id": "installed_version",
+                "label": "Installed version",
+                "model": None,
+                "nullable": True,
+                "primary_key": False,
+                "type": "String",
+            }
+        }
+        assert Blok.fields_description(fields=["installed_version"]) == res
+
+    def test_fields_description_cache(self, rollback_registry):
+        registry = rollback_registry
+        Blok = registry.System.Blok
+        res = {
+            "short_description": {
+                "id": "short_description",
+                "label": "Short description",
                 "model": None,
                 "nullable": True,
                 "primary_key": False,
                 "type": "Function",
-            },
-            "is_sql_model": {
-                "id": "is_sql_model",
-                "label": "Is a SQL model",
-                "model": None,
-                "nullable": True,
-                "primary_key": False,
-                "type": "Boolean",
-            },
-            "name": {
-                "id": "name",
-                "label": "Name",
-                "model": None,
-                "nullable": False,
-                "primary_key": True,
-                "type": "String",
-            },
-            "table": {
-                "id": "table",
-                "label": "Table",
-                "model": None,
-                "nullable": True,
-                "primary_key": False,
-                "type": "String",
-            },
-            "schema": {
-                "id": "schema",
-                "label": "Schema",
-                "model": None,
-                "nullable": True,
-                "primary_key": False,
-                "type": "String",
-            },
-        }
-        assert Model.fields_description() == res
-
-    def test_fields_description_limited_field(self, rollback_registry):
-        registry = rollback_registry
-        Model = registry.System.Model
-        res = {
-            "table": {
-                "id": "table",
-                "label": "Table",
-                "model": None,
-                "nullable": True,
-                "primary_key": False,
-                "type": "String",
             }
         }
-        assert Model.fields_description(fields=["table"]) == res
-
-    def test_fields_description_cache(self, rollback_registry):
-        registry = rollback_registry
-        Model = registry.System.Model
-        Column = registry.System.Column
-        res = {
-            "table": {
-                "id": "table",
-                "label": "Table",
-                "model": None,
-                "nullable": True,
-                "primary_key": False,
-                "type": "String",
-            }
-        }
-        assert Model.fields_description(fields=["table"]) == res
-        column = Column.from_primary_keys(
-            model="Model.System.Model", name="table"
-        )
-        column.label = "Test"
-        assert Model.fields_description(fields=["table"]) == res
-        Model.fire("Update Model", "Model.System.Model")
-        assert Model.fields_description(fields=["table"]) != res
+        assert Blok.fields_description(fields=["short_description"]) == res
+        Blok.loaded_fields["short_description"] = "Test"
+        assert Blok.fields_description(fields=["short_description"]) == res
+        Blok.clear_all_model_caches()
+        assert Blok.fields_description(fields=["short_description"]) != res
 
     def test_to_dict(self, rollback_registry):
         registry = rollback_registry
-        M = registry.System.Model
-        model = M.query().first()
-        assert model.to_dict() == {
-            "name": model.name,
-            "table": model.table,
-            "schema": model.schema,
-            "is_sql_model": model.is_sql_model,
-            "description": model.description,
+        Cache = registry.System.Cache
+        cache = Cache.query().first()
+        assert cache.to_dict() == {
+            "id": cache.id,
+            "method": cache.method,
+            "registry_name": cache.registry_name,
         }
 
     def test_to_dict_on_some_columns(self, rollback_registry):
         registry = rollback_registry
-        M = registry.System.Model
-        model = M.query().first()
-        assert model.to_dict("name", "table") == {
-            "name": model.name,
-            "table": model.table,
+        Blok = registry.System.Blok
+        blok = Blok.query().first()
+        assert blok.to_dict(
+            "name", "installed_version", "short_description"
+        ) == {
+            "name": blok.name,
+            "installed_version": blok.installed_version,
+            "short_description": blok.short_description,
         }
 
     def test_select_sql_statement_with_column(self, rollback_registry):
@@ -205,16 +187,14 @@ class TestCoreSqlBase:
 
     def test_get_hybrid_property_columns(self, rollback_registry):
         registry = rollback_registry
-        Column = registry.System.Column
-        columns = Column.get_hybrid_property_columns()
+        Blok = registry.System.Blok
+        columns = Blok.get_hybrid_property_columns()
         for x in [
             "name",
-            "model",
-            "autoincrement",
-            "foreign_key",
-            "primary_key",
-            "unique",
-            "nullable",
-            "remote_model",
+            "state",
+            "author",
+            "order",
+            "version",
+            "installed_version",
         ]:
             assert x in columns