+++ anyblok/model/cache.py
@@ -17,6 +17,28 @@ class CachePlugin(ModelPluginBase):
 
         super(CachePlugin, self).__init__(registry)
 
+    def insert_in_bases(
+        self, new_base, namespace, properties, transformation_properties
+    ):
+        """Create overload to define the cache from __depends__.
+
+        Because the cache is defined on the depend models and this namespace
+        does not exist in caches dict
+
+        :param new_base: the base to be put on front of all bases
+        :param namespace: the namespace of the model
+        :param properties: the properties declared in the model
+        :param transformation_properties: the properties of the model
+        """
+        for dep in properties["__depends__"]:
+            if dep in self.registry.caches:
+                cache = self.registry.caches.setdefault(namespace, {})
+                for method_name, methods in self.registry.caches[dep].items():
+                    entry = cache.setdefault(method_name, [])
+                    entry.extend(methods)
+
+        return {}
+
     def transform_base_attribute(
         self,
         attr,